+++ internal/tests/convolution_test.go
@@ -4,11 +4,11 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/gomlx/go-xla/pkg/pjrt"
-	. "github.com/gomlx/go-xla/pkg/stablehlo"
-	"github.com/gomlx/go-xla/pkg/types"
-	"github.com/gomlx/go-xla/pkg/types/dtypes"
-	"github.com/gomlx/go-xla/pkg/types/shapes"
+	"github.com/gomlx/compute/dtypes"
+	"github.com/gomlx/go-xla/pjrt"
+	. "github.com/gomlx/go-xla/stablehlo"
+	"github.com/gomlx/go-xla/types"
+	"github.com/gomlx/go-xla/types/shapes"
 )
 
 func TestConvolution(t *testing.T) {